+++ core/testcontainers/core/docker_client.py
@@ -16,10 +16,12 @@ import ipaddress
 import os
 import urllib
 import urllib.parse
+from collections.abc import Iterable
 from typing import Callable, Optional, TypeVar, Union
 
 import docker
 from docker.models.containers import Container, ContainerCollection
+from docker.models.images import Image, ImageCollection
 from typing_extensions import ParamSpec
 
 from testcontainers.core.config import testcontainers_config as c
@@ -34,6 +36,14 @@ _T = TypeVar("_T")
 
 def _wrapped_container_collection(function: Callable[_P, _T]) -> Callable[_P, _T]:
     @ft.wraps(ContainerCollection.run)
+    def wrapper(*args: _P.args, **kwargs: _P.kwargs) -> _T:
+        return function(*args, **kwargs)
+
+    return wrapper
+
+
+def _wrapped_image_collection(function: Callable[_P, _T]) -> Callable[_P, _T]:
+    @ft.wraps(ImageCollection.build)
     def wrapper(*args: _P.args, **kwargs: _P.kwargs) -> _T:
         return function(*args, **kwargs)
 
@@ -93,6 +103,17 @@ class DockerClient:
             **kwargs,
         )
         return container
+
+    @_wrapped_image_collection
+    def build(self, path: str, tag: str, rm: bool = True, **kwargs) -> tuple[Image, Iterable[dict]]:
+        """
+        Build a Docker image from a directory containing the Dockerfile.
+
+        :return: A tuple containing the image object and the build logs.
+        """
+        image_object, image_logs = self.client.images.build(path=path, tag=tag, rm=rm, **kwargs)
+
+        return image_object, image_logs
 
     def find_host_network(self) -> Optional[str]:
         """
@@ -166,18 +187,14 @@ class DockerClient:
         """
         Get the hostname or ip address of the docker host.
         """
-        # https://github.com/testcontainers/testcontainers-go/blob/dd76d1e39c654433a3d80429690d07abcec04424/docker.go#L644
-        # if os env TC_HOST is set, use it
-        host = os.environ.get("TC_HOST")
-        if not host:
-            host = os.environ.get("TESTCONTAINERS_HOST_OVERRIDE")
+        host = c.tc_host_override
         if host:
             return host
         try:
             url = urllib.parse.urlparse(self.client.api.base_url)
 
         except ValueError:
-            return None
+            return "localhost"
         if "http" in url.scheme or "tcp" in url.scheme:
             return url.hostname
         if inside_container() and ("unix" in url.scheme or "npipe" in url.scheme):